Add filecheck output to tests
