+++ test/UnparserTests/scanf_sanitize.c
@@ -1,0 +1,24 @@
+// RUN: unparse %s | clang-format | FileCheck %s --color
+
+#include <stdio.h>
+
+void sanitize(int *x) {
+}
+
+int main(void) {
+        int x;
+
+        scanf("%d", &x);
+        sanitize(&x);
+
+        return 0;
+}
+
+// CHECK: void sanitize(int *x) {
+// CHECK: }
+// CHECK: int main(void) {
+// CHECK:         int x;
+// CHECK:         scanf("%d", &x);
+// CHECK:         sanitize(&x);
+// CHECK:         return 0;
+// CHECK: }
