--- test/assgn_p_san.c
@@ -1,21 +1,0 @@
-#include <stdio.h>
-
-void sanitize(int *x) {
-        if (*x != 0) {
-                *x = 0;
-        } else {
-        }
-}
-
-int main(void) {
-        int x;
-        int y;
-        scanf("%d", &x);
-
-        y = x;
-        y = 5;
-
-        sanitize(&x);
-
-        return 0;
-}
