--- test/assgn_san.c
@@ -1,19 +1,0 @@
-#include <stdio.h>
-
-void sanitize(int *x) {
-        if (*x != 0) {
-                *x = 0;
-        } else {
-        }
-}
-
-int main(void) {
-        int x;
-        int y;
-        scanf("%d", &x);
-
-        y = x;
-        y = 5;
-        x = 3;
-        return 0;
-}
