--- test/assign.c
@@ -1,12 +1,0 @@
-#include <stdio.h>
-
-void sanitize(int *x) {
-}
-
-int main(void) {
-        int x;
-
-        x = 1;
-
-        return 0;
-}
