--- test/bin_expr.c
@@ -1,13 +1,0 @@
-#include <stdio.h>
-
-void sanitize(int *x) {
-        return;
-}
-
-int main(void) {
-        int x;
-
-        x = 1 + 2 * 3 - 4;
-
-        return 0;
-}
