--- test/decls.c
@@ -1,13 +1,0 @@
-#include <stdio.h>
-
-void sanitize(int *x) {
-}
-
-int main(void) {
-        int a;
-        int b;
-        int c;
-        int d;
-        int e;
-        return 0;
-}
