--- test/fix_flow_t.c
@@ -1,19 +1,0 @@
-#include <stdio.h>
-
-void sanitize(int *x) {
-        if (*x != 0) {
-                *x = 0;
-        } else {
-        }
-}
-
-int main(void) {
-        int x;
-        int y;
-        scanf("%d", &x);
-        y = x;
-        sanitize(&x);
-        sanitize(&y);
-
-        return 0;
-}
