--- test/returns.c
@@ -1,11 +1,0 @@
-#include <stdio.h>
-
-void sanitize(int *x) {
-        return;
-        return;
-}
-
-int main(void) {
-        return 1;
-        return 0;
-}
