--- test/san_only.c
@@ -1,16 +1,0 @@
-#include <stdio.h>
-
-void sanitize(int *x) {
-        if (*x != 0) {
-                *x = 0;
-        } else {
-        }
-}
-
-int main(void) {
-        int x;
-
-        sanitize(&x);
-
-        return 0;
-}
