--- test/sanitize_main.c
@@ -1,8 +1,0 @@
-#include <stdio.h>
-
-void sanitize(int *x) {
-}
-
-int main(void) {
-        return 0;
-}
