--- test/skip.c
@@ -1,13 +1,0 @@
-#include <stdio.h>
-
-void sanitize(int *x) {
-}
-
-int main(void) {
-        ;
-        ;
-        ;
-        ;
-        ;
-        return 0;
-}
