--- test/tainted.c
@@ -1,11 +1,0 @@
-#include <stdio.h>
-
-void sanitize(int *x) {
-}
-
-int main(void) {
-        int x;
-        scanf("%d", &x);
-
-        return 0;
-}
